refactor interval based into own package
