+++ aeon/classification/interval_based/_interval_pipelines.py
@@ -11,8 +11,7 @@ from sklearn.ensemble import RandomForestClassifier
 
 from aeon.base._base import _clone_estimator
 from aeon.classification.base import BaseClassifier
-from aeon.transformations.collection import SupervisedIntervals
-from aeon.transformations.collection.random_intervals import RandomIntervals
+from aeon.transformations.collection.interval_based import SupervisedIntervals, RandomIntervals
 
 
 class RandomIntervalClassifier(BaseClassifier):
