+++ aeon/transformations/collection/__init__.py
@@ -16,7 +16,6 @@ __all__ = [
     "MatrixProfile",
     "PaddingTransformer",
     "PeriodogramTransformer",
-    "RandomIntervals",
     "Tabularizer",
     "TimeBinner",
     "IntervalSegmenter",
@@ -24,7 +23,6 @@ __all__ = [
     "SlidingWindowSegmenter",
     "SlopeTransformer",
     "SevenNumberSummaryTransformer",
-    "SupervisedIntervals",
     "TimeSeriesScaler",
     "TruncationTransformer",
     "TSFreshFeatureExtractor",
@@ -48,7 +46,6 @@ from aeon.transformations.collection.interpolate import TSInterpolator
 from aeon.transformations.collection.matrix_profile import MatrixProfile
 from aeon.transformations.collection.pad import PaddingTransformer
 from aeon.transformations.collection.periodogram import PeriodogramTransformer
-from aeon.transformations.collection.random_intervals import RandomIntervals
 from aeon.transformations.collection.reduce import Tabularizer, TimeBinner
 from aeon.transformations.collection.scaler import TimeSeriesScaler
 from aeon.transformations.collection.segment import (
@@ -58,7 +55,6 @@ from aeon.transformations.collection.segment import (
 )
 from aeon.transformations.collection.slope import SlopeTransformer
 from aeon.transformations.collection.summary import SevenNumberSummaryTransformer
-from aeon.transformations.collection.supervised_intervals import SupervisedIntervals
 from aeon.transformations.collection.truncate import TruncationTransformer
 from aeon.transformations.collection.tsfresh import (
     TSFreshFeatureExtractor,
