+++ aeon/transformations/collection/interval_based/__init__.py
@@ -1,0 +1,9 @@
+"""Interval based collection transformations."""
+
+__all__ = [
+    "RandomIntervals",
+    "SupervisedIntervals",
+]
+from aeon.transformations.collection.interval_based._random_intervals import RandomIntervals
+from aeon.transformations.collection.interval_based._supervised_intervals import \
+    SupervisedIntervals
