+++ aeon/transformations/collection/interval_based/_random_intervals.py
