+++ aeon/transformations/collection/interval_based/_supervised_intervals.py
