+++ aeon/transformations/collection/interval_based/tests/test_intervals.py
@@ -1,8 +1,7 @@
 """Interval extraction test code."""
 
 from aeon.transformations.collection import Catch22, SevenNumberSummaryTransformer
-from aeon.transformations.collection.random_intervals import RandomIntervals
-from aeon.transformations.collection.supervised_intervals import SupervisedIntervals
+from aeon.transformations.collection.interval_based import RandomIntervals, SupervisedIntervals
 from aeon.utils._testing.collection import make_3d_test_data
 from aeon.utils.numba.stats import row_mean, row_median
