+++ aeon/transformations/tests/test_transformer_reproduction.py
@@ -4,9 +4,8 @@ import numpy as np
 from sklearn.utils._testing import set_random_state
 
 from aeon.datasets import load_basic_motions, load_unit_test
-from aeon.transformations.collection.random_intervals import RandomIntervals
+from aeon.transformations.collection.interval_based import RandomIntervals, SupervisedIntervals
 from aeon.transformations.collection.shapelet_based import RandomShapeletTransform
-from aeon.transformations.collection.supervised_intervals import SupervisedIntervals
 
 
 def _reproduce_transform_unit_test(estimator):
